services/nerve-center/requirements.txt
fastapi
uvicorn
azure-identity
azure-servicebus
azure-cosmos
requests
